+++ packages/integration-sdk-testing/src/config.ts
@@ -1,0 +1,13 @@
+import {
+  IntegrationInstanceConfig,
+  IntegrationInvocationConfig,
+} from '@jupiterone/integration-sdk-core';
+
+export interface StepTestConfig<
+  TInvocationConfig extends IntegrationInvocationConfig = IntegrationInvocationConfig,
+  TInstanceConfig extends IntegrationInstanceConfig = IntegrationInstanceConfig
+> {
+  stepId: string;
+  invocationConfig: TInvocationConfig;
+  instanceConfig: TInstanceConfig;
+}