+++ packages/integration-sdk-testing/src/index.ts
@@ -1,5 +1,7 @@
 export * from './context';
+export * from './config';
 export * from './executeStepWithDependencies';
+export * from './filterGraphObjects';
 export * from './logger';
 export * from './recording';
 export * from './jobState';